locale.c: Use widest int type for parameter
nl_langinfo() is a libc function, specified by POSIX, that returns
information about some particular element of a locale.  Which element is
desired is passed to nl_langinfo() via a parameter with a typedef of
'nl_item'.  What that typedef actually is is opaque to perl.  On my
Linux glibc, it is an 'int', but it could just as easily be an enum.
POSIX lists all the elements that nl_langinfo() should handle, but some
platforms have extensions beyond those.  Perl aims to handle any
somewhat prominent such extension, which means it needs fallbacks if the
particular item isn't available on a given platform.

emulate_langinfo() is used to handle such fallbacks.  If an element
isn't available on a platform, perl has to create an element number that
doesn't clash with any nl_item one.  perl also has to cope with
platforms that don't have nl_langinfo() at all (Win32 is one).  In this
case it has to create an 'nl_item' typedef.  We use 'int'.

So emulate_langinfo() has to be able to handle an 'int' or whatever the
system specifies an nl_item may be.  This could be a 'long' or a
'long long'.  The only way to do that is to make the parameter as wide
as the the platform permits.  That is what this commit does.

This commit also casts the item to the widest possible value with
corresponding %j format for display in a Debug statement in the
sv_langinfo_i() function.