regcomp_internal.h - handle -Uusedl in RExC_state_t
Under -Uusedl there were two problems: the first was that we were
conditionally adding fields based on whether DEBUGGING was true, but
this does not work properly when do a static build as we use the same
data structure definition for DEBUGGING and non-DEBUGGING builds as we
compile both into the same library. So in this case we just always
include the additional fields.

We probably could even just ALWAYS include them, as we normally only
have one of these structs around so the unused storage for these fields
would not really be a burden.

The second problem was that the debugging fields were not at the end of
the structure. This meant that code compiled under DEBUGGING and not
would access the same memory for different purposes.

This patch makes it so we include these fields if we are building under
DEBUGGING, OR if we are doing a static build. It also moves these fields
to the end so that there is no ambiguity about which memory is used
for what.