t/op/sub.t: silence new warning
A 7-year old test has recently started emitting this warning:

    Lexical subroutine &b masks previously declared package subroutine

because a new warning has been added to core.

It's not 100% clear to me whether the test itself was *relying* on a
package subroutine having the same name or not, but given that it was a
test which crashed rather than had well-defined "bad" behaviour", I
decided it was best to leave the test exactly as-is, and just silence
the warning.