reentr.pl: Increase strerror_r buffer size
We have seen cases where a libc strerror() buffer size of 80 is not
enough.  See https://github.com/Perl/perl5/issues/21992.

One might think our current size for strerror_r() of 256 would handle
any situation, but in looking around for this, I saw that glibc bothers
to document that theirs is 1024.  So this commit bumps it to that.