allow multiple use v5.x declarations in some cases where it's harmless
We want to disallow using multiple "use v5.x" declarations in the same
scope, but there are some cases we know to be harmless.

For versions before 5.10, the use declaration was only a version check
with no side effects. If there was a previous version declarations that
was below 5.10, we can safely ignore it without a warning or error.