Don't run File::Find tests in parallel
We were seeing intermittent smoke failures for find.t and taint.t, with
errors in the logs like

    Can't cd to .. from ./FF_find_t_RKdkBE/for_find/fb: Stale file handle

where the temp dir it was complaining about is one actually created by
the *other* test file.

So it seems that it's best not to run these two tests in parallel.
I suppose a more complete solution would be to have two separate
permanent test dirs with two sets of test files for the two test
scripts to play with, but that sounds like a lot of work.

In the meantime, lets see if the problem goes away with this quick fix.