locale_header.h: Add explicit #define
The previous commit made this a generated file.  Prior to this commit
one #define was omitted because it would be interpreted as 0 anyway, and
it was a pain to add to each of the entries.  But now, it is easy to
add, and tends to be clearer to have an explicit #define.