win32.c: suppress -Wcast-function-type warnings
This appears to be the only place these warnings are produced even
though win32/win32.c contains several other function type casts.

None of the non-warning casts use the "void (*)(void)" type that is
documented to suppress the warning, and adding such a cast to
the lines warning seems less safe to me, cast from a stdcall
function pointer to a cdecl function pointer and back.

So, just suppress the warning where I saw it.