Extutils::ParseXS *.pm: add many code comments
To aid the understanding of this module and its sub-modules:

 - add lots of code comments
 - add blank lines
 - reformat and/or line-wrap a few long code lines

There should be no functional changes.

In particular, the line count of ParseXS.pm is increased by about 60%
with this commit.

I've tried to consistently use the word 'emit' rather than 'print' or
'output' in comments about the code that gets generated and ends up in
the .c file. At the moment most of this code code is indeed just
immediately printed to STDOUT, but in the longer term I would like to
separate out code generation and output stages.