Keep trailing comma in PL_magic_vtable(_name)s definitions
C99 allows trailing commas in on struct and table initialisers, so get
rid of the code to avoid it in PL_magic_vtables and adjust the _names
code to match.

We don't add it to the enum, since magic_vtable_max is always the last
element, so there won't be any diff noise from adding new elements
after it.