Keep using undocumented gcc locale feature
Since 5.39.7, this feature has been enabled experimentally by default.
It is similar to querylocale(), found on *BSD systems.  That function
returns the current locale, and was created because this ability is
often required, and POSIX neglected to define anything for this purpose.
I'm told that the next POSIX Standard update will include a function for
this purpose.

I have been using the gcc feature successfully for years in my local
work spaces; we started this experiment to see if it worked generally.
And there have been zero reports of problems with it.

The experiment was designed to refuse to compile in 5.39.10, so as to
make sure it didn't slip into 5.40 without it being explicitly decided
to do so.  Now, the final steps in publishing 5.39.9 include bumping the
release to 5.39.10, and those can't be done with the code as-is.

This commit declares the experiment a success, and enables this feature
non-experimentally as the default, with the ability to turn it off by
passing
    -Accflags=-DNO_USE_NL_LOCALE_NAME
to Configure.