Porting/sync-with-cpan - don't die if the Changes entry detection fails.
We use a heuristic to detect Changes file entries to automate producing
the correct commit message when we update a distribution. If the Changes
file format changes in various the heuristics might break. We don't want
to break the upgrade process for this. We should just fallback to the
old manual commit message process instead.