Change name of non-public function
Any name beginning with an underscore is reserved for the C
implementation itself.  It is undefined behavior to use one, though the
only practical consequences are a possible naming conflict, which is
very unlikely to happen.

But we are gradually fixing these when they otherwise get touched, as
this series of commits does with "_is_in_locale_category".  This is the
short name defined in embed.h.  This commit changes the leading
underscore to a trailing, which is legal for us to use.