Free up HINT bit
There are two locale-related hint bits.  This commit combines them into
one.  There is no slowdown in checking if one of the forms is in effect,
but distinguishing a plain 'use locale' from the other form is slightly
slower.

This works because $H{locale} already is 0 for one form; non-zero for
the other.