Store PL_prevailing_version in the frozen hinthash around a string-eval op
We need to save the value of PL_prevailing_version at the time the eval
op was compiled, so it can be put in place during the running code.

Ideally we'd do something more robust, like change the OP_ENTERVAL op
class into UNOP_AUX, so that the aux vector can store additional
information like the version number and perhaps the frozen hints hash.

In practice it is far too close to the 5.40 release to contemplate such
a change now, so this is a less intrusive but hackier change to achieve
the same aim.

See also
  https://github.com/Perl/perl5/issues/22121