Fcntl: remove pointless S_ISENFMT function
S_ISENFMT($mode) has no equivalent C macro. It was implemented (in XS)
as the equivalent of `($mode & S_IFMT) == S_ENFMT`, which is always
false because S_ENFMT is not a file type and thus not covered by the
S_IFMT mask.

(See commits ca6e1c2, 0e6f150, 0f68039, fb59364, 87eca6e for the history
of this function, which has always been broken in various ways ever
since it was added in 2000.)

Fixes #22190.