S_new_SV: args unused, static inline & defined, rename with Perl_ prefix
When sv_inline.h was created in
https://github.com/Perl/perl5/commit/75acd14e43f2ffb698fc7032498f31095b56adb5
and a number of things moved into it, the S_new_SV debugging function should
have been made PERL_STATIC_INLINE and given the Perl_ prefix. This commit
now does those things. It also marks the arguments to Perl_new_SV as
PERL_UNUSED_ARG, reducing warnings on some builds.

Additionally, now that we can use inline functions, the new_SV() macro is
now just a call to this inline function, rather than being that only on
DEBUG_LEAKING_SCALARS builds and a multi-line macro the rest of the time.