make (*THEN) work - first draft
Currently, a global flag (RExC_seen |= REG_CUTGROUP_SEEN;) is set in regcomp.c when (*THEN) is encountered. Then in S_regmatch() and when encountering a BRANCH node, a choice point is puhsed with PUSH_YES_STATE_GOTO if the cutgroup flag is set, otherwise a normal choice point is pushed with PUSH_STATE_GOTO. Since this flag is global to the regex, every BRANCH node of the regex will push will a "yes state" if the regex contain a single (*THEN), instead of BRANCH nodes pushing regular choice points.

Given that "sayNO" (e.g. simple backtracking) performs a cut if a "yes state" (choice point and cut point) remains and if no_final is true (which it is because backtracking after having encountered (*THEN), to the choice point pushed by (*THEN) causes the regex engine to execute the op CUTGOUP_next_fail which assigns no_final to 1) and that every alternation is also a new cut point, the cut operation has the effect to perform a simple backtracking operation as far as alternations are concerned, because every alternation pushes a choice point/cut point (e.g "yes state"). Normally, only the BRANCH node of the alternation directly containing (*THEN) should push a "yes state".

One solution is to mark the appropriate BRANCH as being the outtermost alternation of a (*THEN) group and check this mark/flag inside case BRANCH (which it already does currently) and push a new choice point or a choice point+cut point accordingly with PUSH_YES_STATE_GOTO or respectively PUSH_STATE_GOTO. An alternative solution is to create a distinct BRANCH_cutgroup regnode type which would always push a choice point+cut point while the BRANCH regnode would always push a regular choice point.

In order to mark BRANCH nodes as containing a (*THEN), a field (has_cutgroup) in the regex parser and compiler context (RExC_state_t) is set to 1 in regbranch() when encountering (*THEN). This field is looked up by S_reg() (regbranch()'s caller) when creating the BRANCH regnode and set the regnode_head data.u_8.flags field of this new regnode accordingly. This field (has_cutgroup) is first saved before calling regbranch(), then set to 0, then call regbranch() is made. After regbranch() returned and after the has_cutgroup field has been checked when creating the regnode, it is restored to its previous value in order to not forget having seen a previous (*THEN) group in a situation like: / (?: a (*THEN) (?: b (*THEN) | c) | d) /.

!!! This commit doesn't attempt to fix the logic of TRIE regnodes concerning (*THEN) !!!